feat(liveman): use cargo bin
