+++ liveman/Cargo.toml
@@ -6,6 +6,9 @@ edition.workspace = true
 license.workspace = true
 repository.workspace = true
 
+[lib]
+crate-type = ["lib"]
+
 [dependencies]
 liveion = { path = "../liveion", optional = true }
 net4mqtt = { path = "../libs/net4mqtt", optional = true }
@@ -14,7 +17,6 @@ api = { path = "../libs/api" }
 auth = { path = "../libs/auth" }
 http-log = { path = "../libs/http-log" }
 signal = { path = "../libs/signal" }
-utils = { path = "../libs/utils" }
 
 anyhow = { workspace = true, features = ["backtrace"] }
 clap = { workspace = true, features = ["derive"] }
