+++ liveman/src/config.rs
@@ -165,7 +165,7 @@ impl Default for CheckReforwardTickTime {
 }
 
 impl Config {
-    pub(crate) fn parse(path: Option<String>) -> Self {
+    pub fn parse(path: Option<String>) -> Self {
         let result = fs::read_to_string(path.unwrap_or(String::from("liveman.toml")))
             .or(fs::read_to_string("/etc/live777/liveman.toml"))
             .unwrap_or("".to_string());
